+++ packages/terra-functional-testing/src/services/wdio-visual-regression-service/VisualRegressionLauncher.js
@@ -7,12 +7,15 @@ const getTerraFormFactor = require('./modules/getTerraFormFactor');
 
 class VisualRegressionLauncher {
   /**
-   * @param {Object} options - Service configuration options.
-   * @param {Object} options.locale - The locale being tested.
-   * @param {Object} options.theme - The theme being tested.
+   * Service constructor.
+   * @param {Object} _options - The options specific to this service.
+   * @param {Object} _capabilities - The list of capabilities details.
+   * @param {Object} config - The object containing the wdio configuration and options defined in the terra-cli test runner.
    */
-  constructor(options) {
-    this.compare = new LocalCompare(options);
+  constructor(_options, _capabilities, config = {}) {
+    const { launcherOptions } = config;
+
+    this.compare = new LocalCompare(launcherOptions);
     this.context = null;
     this.currentSuite = null;
     this.currentTest = null;
@@ -21,7 +24,7 @@ class VisualRegressionLauncher {
   /**
    * Gets executed before test execution begins. At this point you can access
    * all global variables, such as `browser`.
-   * @param {object} capabilities - desiredCapabilities
+   * @param {Object} capabilities - desiredCapabilities
    * @param {[type]} specs
    * @return null
    */
@@ -68,20 +71,21 @@ class VisualRegressionLauncher {
    * Command wrapper to setup the command with the correct context values defined from the global
    * webdriver.IO WebDriver instance.
    *
-   * @param {object} browser - The global webdriver.IO WebDriver instance.
+   * @param {Object} browser - The global webdriver.IO WebDriver instance.
    * @param {function} command - The test command that should be executed.
    */
   wrapCommand(browser, command) {
     /**
      * The wrapped command with access to the global webdriver.IO WebDriver instance.
      *
-     * @param {String} elementSelector - The css selector of the element that should be captured in the screenshot.
+     * @param {string} elementSelector - The css selector of the element that should be captured in the screenshot.
      * @param {Object=} options - The screenshot capturing and comparison options.
-     * @param {String[]} options.hide - The list of elements to set opacity 0 on to 'hide' from the dom when capturing the screenshot.
-     * @param {String[]} options.remove - The list of elements to set display: none on to 'remove' from dom when capturing the screenshot.
-     * @param {String} options.ignoreComparison - The image comparison algorithm to use when processing the screenshot comparison.
-     * @param {Number} options.mismatchTolerance - The acceptable mismatch tolerance the screenshot can have when processing the screenshot comparison.
-     * @param {String} options.name - The name of the screenshot.
+     * @param {string[]} options.hide - The list of elements to set opacity 0 on to 'hide' from the dom when capturing the screenshot.
+     * @param {string[]} options.remove - The list of elements to set display: none on to 'remove' from dom when capturing the screenshot.
+     * @param {string} options.ignoreComparison - The image comparison algorithm to use when processing the screenshot comparison.
+     * @param {number} options.mismatchTolerance - The acceptable mismatch tolerance the screenshot can have when processing the screenshot comparison.
+     * @param {string} options.name - The name of the screenshot.
+     * @param {boolean} options.updateScreenshots - Whether or not to automatically update all reference screenshots with the latest screenshots.
      * @returns {Object} - The screenshot comparison results returned as { misMatchPercentage: Number, isSameDimensions: Boolean, getImageDataUrl: function }.
      */
     return async function wrappedScreenshotCommand(elementSelector, options) {