+++ packages/terra-functional-testing/src/services/wdio-visual-regression-service/modules/makeElementScreenshot.js
@@ -14,11 +14,11 @@ const logger = new Logger('[wdio-visual-regression-service:makeDocumentScreensho
  *
  * @alias browser.checkElement
  * @param {string=} fileName - The filename to use to save the screenshot.
- * @param {String} elementSelector - The css selector of the element that should be captured in the screenshot.
+ * @param {string} elementSelector - The css selector of the element that should be captured in the screenshot.
  * @param {Object=} options - The screenshot capturing and comparison options.
- * @param {String[]} options.hide - The list of elements to set opacity 0 on to 'hide' from the dom when capturing the screenshot.
- * @param {String[]} options.remove - The list of elements to set display: none on to 'remove' from dom when capturing the screenshot.
- * @returns {String} - The base64 string of the screenshot image that was captured.
+ * @param {string[]} options.hide - The list of elements to set opacity 0 on to 'hide' from the dom when capturing the screenshot.
+ * @param {string[]} options.remove - The list of elements to set display: none on to 'remove' from dom when capturing the screenshot.
+ * @returns {string} - The base64 string of the screenshot image that was captured.
  */
 async function makeElementScreenshot(browser, elementSelector, options = {}) {
   logger.verbose('start element screenshot');