+++ packages/terra-functional-testing/src/services/wdio-visual-regression-service/scripts/scroll.js
@@ -1,8 +1,8 @@
 /**
  * Script to scroll to a particular set of coordinates in the document.
  *
- * @param {Number} x - The pixel along the horizontal axis of the document.
- * @param {Number} y - The pixel along the vertical axis of the document.
+ * @param {number} x - The pixel along the horizontal axis of the document.
+ * @param {number} y - The pixel along the vertical axis of the document.
  * @returns {undefined}
  */
 function scroll(x, y) {