+++ packages/terra-functional-testing/tests/jest/services/wdio-terra-service.test.js
@@ -13,6 +13,18 @@ const mockFindElement = jest.fn().mockImplementation(() => element);
 const TerraService = () => { };
 const serviceOptions = { formFactor: 'huge' };
 
+const config = {
+  serviceOptions: {
+    selector: 'mock-selector',
+  },
+  launcherOptions: {
+    theme: 'mock-theme',
+    formFactor: 'huge',
+  },
+};
+
+const capabilities = { browserName: 'chrome' };
+
 global.browser = {
   $: mockFindElement,
   options: {
@@ -26,20 +38,17 @@ global.browser = {
 global.Terra = {};
 
 describe('WDIO Terra Service', () => {
-  it('should setup the global terra validates accessibility command', () => {
-    const service = new WdioTerraService();
+  it('should setup the global terra object in before hook', () => {
+    const service = new WdioTerraService({}, {}, config);
 
-    service.before({ browserName: 'chrome' });
+    service.before(capabilities);
 
     expect(global.Terra.validates.accessibility).toBeDefined();
-  });
-
-  it('should setup the global terra validates element command', () => {
-    const service = new WdioTerraService();
-
-    service.before({ browserName: 'chrome' });
-
     expect(global.Terra.validates.element).toBeDefined();
+    expect(global.Terra.validates.screenshot).toBeDefined();
+    expect(global.Terra.hideInputCaret).toBeDefined();
+    expect(global.Terra.setApplicationLocale).toBeDefined();
+    expect(setViewport).toHaveBeenCalledWith(service.serviceOptions.formFactor);
   });
 
   it('should setup the global terra axe configuration', () => {
@@ -52,9 +61,9 @@ describe('WDIO Terra Service', () => {
   });
 
   it('should disable the axe color contrast rule for lowlight theme', () => {
-    const service = new WdioTerraService({ theme: 'clinical-lowlight-theme' });
+    const service = new WdioTerraService({}, {}, { launcherOptions: { theme: 'clinical-lowlight-theme' } });
 
-    service.before({ browserName: 'chrome' });
+    service.before(capabilities);
 
     const rules = {
       'scrollable-region-focusable': { enabled: false },
@@ -62,23 +71,6 @@ describe('WDIO Terra Service', () => {
     };
 
     expect(global.Terra.axe).toEqual({ rules });
-  });
-
-  it('should set the expect command as a global api', () => {
-    const service = new WdioTerraService();
-
-    service.before({ browserName: 'chrome' });
-
-    expect(expect).toBeDefined();
-  });
-
-  it('should set viewport helper commands as as global api', () => {
-    const service = new WdioTerraService({ formFactor: 'huge' });
-
-    service.before({ browserName: 'chrome' });
-
-    expect(setViewport).toBeCalled();
-    expect(global.Terra.hideInputCaret).toBeDefined();
   });
 
   it('should wait for browser interaction for IE', () => {
@@ -92,7 +84,7 @@ describe('WDIO Terra Service', () => {
   it('should hide input caret after command', () => {
     const service = new WdioTerraService();
 
-    service.before({ browserName: 'chrome' });
+    service.before(capabilities);
 
     const mockHideInputCaret = jest.fn();
     global.Terra.hideInputCaret = mockHideInputCaret;
@@ -103,51 +95,51 @@ describe('WDIO Terra Service', () => {
     expect(mockFindElement).toHaveBeenCalledWith('[data-terra-test-loading]');
   });
 
-  it('should define commands in beforeSession with empty config', () => {
-    const service = new WdioTerraService({ formFactor: 'huge' });
+  it('should define commands in beforeSession with no serviceOptions', () => {
+    const service = new WdioTerraService();
     const expectedServiceOptions = {
-      formFactor: 'huge',
       selector: '[data-terra-test-content] *:first-child',
       theme: 'terra-default-theme',
     };
 
-    service.beforeSession({});
+    service.beforeSession();
     expect(global.Terra.describeViewports).toBeDefined();
     expect(global.Terra.viewports).toBeDefined();
     expect(global.Terra.serviceOptions).toEqual(expectedServiceOptions);
   });
 
-  it('should define commands in beforeSession', () => {
-    const service = new WdioTerraService({ formFactor: 'huge' });
-    const config = {
-      serviceOptions: {
-        selector: 'mock-selector',
-      },
+  it('should set service options with empty config', () => {
+    const service = new WdioTerraService();
+    const expectedServiceOptions = {
+      selector: '[data-terra-test-content] *:first-child',
+      theme: 'terra-default-theme',
     };
+
+    expect(service.serviceOptions).toEqual(expectedServiceOptions);
+  });
+
+  it('should set service options with populated config', () => {
+    const service = new WdioTerraService({}, {}, config);
+    const expectedServiceOptions = {
+      formFactor: 'huge',
+      selector: 'mock-selector',
+      theme: 'mock-theme',
+    };
+
+    expect(service.serviceOptions).toEqual(expectedServiceOptions);
+  });
+
+  it('should define all commands', () => {
+    const service = new WdioTerraService({}, {}, config);
 
     const expectedServiceOptions = {
       formFactor: 'huge',
       selector: 'mock-selector',
-      theme: 'terra-default-theme',
+      theme: 'mock-theme',
     };
 
-    service.beforeSession(config);
-    expect(global.Terra.describeViewports).toBeDefined();
-    expect(global.Terra.viewports).toBeDefined();
-    expect(global.Terra.serviceOptions).toEqual(expectedServiceOptions);
-  });
-
-  it('should define all commands', () => {
-    const service = new WdioTerraService({ formFactor: 'large' });
-
-    const expectedServiceOptions = {
-      formFactor: 'large',
-      selector: '[data-terra-test-content] *:first-child',
-      theme: 'terra-default-theme',
-    };
-
-    service.beforeSession({});
-    service.before({ browserName: 'chrome' });
+    service.beforeSession();
+    service.before(capabilities);
 
     expect(setViewport).toBeCalled();
     expect(global.Terra.viewports).toBeDefined();