+++ packages/terra-functional-testing/CHANGELOG.md
@@ -2,4 +2,11 @@
 
 ## Unreleased
 
+## 1.0.1 - (March 1, 2021)
+
+* Fixed
+  * Correctly pass theme as `defaultTheme` to webpack-config-terra to run tests in the correct theme.
+
+## 1.0.0 - (February 25, 2021)
+
 * Initial stable release