+++ packages/terra-open-source-scripts/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 1.7.0 - (February 25, 2021)
+
+* Changed
+  * Minor dependency version bump
+
+## 1.6.0 - (February 8, 2021)
+
+* Changed
+  * Minor dependency version bump
+
 ## 1.5.0 - (January 26, 2021)
 
 * Changed