+++ packages/terra-toolkit-docs/CHANGELOG.md
@@ -2,6 +2,23 @@
 
 ## Unreleased
 
+## 1.8.1 - (March 1, 2021)
+
+* Changed
+  * Add @cerner/terra-functional-testing to dependencies
+  * Update doc from `getCssProperty` to `getCSSProperty`
+  
+
+## 1.8.0 - (February 25, 2021)
+
+* Changed
+  * Minor dependency version bump
+
+## 1.7.0 - (February 16, 2021)
+
+* Added
+  * Added documentation for eslint-config-terra
+
 ## 1.6.0 - (January 27, 2021)
 
 * Changed